# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/events": {
      "get": {
        "operationId": "EventsController_findAll",
        "summary": "List all events",
        "description": "Returns all events. Public clients should render only approved events.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. Accepted values: admin | community_manager | organizer | moderator | user. GET / POST require any valid role. PATCH requires community_manager, organizer, or admin. DELETE requires admin.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "community_manager",
                "organizer",
                "moderator",
                "user"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter events by status",
            "schema": {
              "enum": [
                "pending",
                "approved",
                "rejected",
                "cancelled"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of event records",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventDto"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Missing or invalid x-role header"
          }
        },
        "tags": [
          "events"
        ]
      },
      "post": {
        "operationId": "EventsController_create",
        "summary": "Create an event",
        "description": "Creates an event request. Users submit pending events; Community Managers/Admins can publish approved events.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. Accepted values: admin | community_manager | organizer | moderator | user. GET / POST require any valid role. PATCH requires community_manager, organizer, or admin. DELETE requires admin.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "community_manager",
                "organizer",
                "moderator",
                "user"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Event creation payload",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The newly created event",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDto"
                }
              }
            }
          },
          "403": {
            "description": "Missing or invalid x-role header"
          }
        },
        "tags": [
          "events"
        ]
      }
    },
    "/api/events/{id}": {
      "get": {
        "operationId": "EventsController_findOne",
        "summary": "Get a single event by ID",
        "description": "Fetches one event by its numeric ID.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. Accepted values: admin | community_manager | organizer | moderator | user. GET / POST require any valid role. PATCH requires community_manager, organizer, or admin. DELETE requires admin.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "community_manager",
                "organizer",
                "moderator",
                "user"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Numeric event ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The matching event record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDto"
                }
              }
            }
          },
          "403": {
            "description": "Missing or invalid x-role header"
          },
          "404": {
            "description": "Event not found"
          }
        },
        "tags": [
          "events"
        ]
      },
      "patch": {
        "operationId": "EventsController_update",
        "summary": "Update an event",
        "description": "Partially updates an event and supports approve/reject status changes. Requires community_manager, organizer, or admin.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. Accepted values: admin | community_manager | organizer | moderator | user. GET / POST require any valid role. PATCH requires community_manager, organizer, or admin. DELETE requires admin.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "community_manager",
                "organizer",
                "moderator",
                "user"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Numeric event ID to update",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Fields to update (all optional)",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated event record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDto"
                }
              }
            }
          },
          "403": {
            "description": "Only community_manager, organizer, or admin can update events"
          },
          "404": {
            "description": "Event not found"
          }
        },
        "tags": [
          "events"
        ]
      },
      "delete": {
        "operationId": "EventsController_remove",
        "summary": "Delete an event",
        "description": "Permanently deletes an event. Requires x-role: admin.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. Accepted values: admin | community_manager | organizer | moderator | user. GET / POST require any valid role. PATCH requires community_manager, organizer, or admin. DELETE requires admin.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "community_manager",
                "organizer",
                "moderator",
                "user"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Numeric event ID to delete",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deletion confirmation",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Event 2 deleted"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Only admin can delete events"
          },
          "404": {
            "description": "Event not found"
          }
        },
        "tags": [
          "events"
        ]
      }
    },
...
    "/api/memberships": {
      "get": {
        "operationId": "MembershipsController_findAll",
        "summary": "List all memberships, optionally filtered by communityId or userId",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC: admin | community_manager | moderator | user",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "communityId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": [
                    {
                      "id": 1,
                      "userId": 3,
                      "communityId": 1,
                      "joinedAt": "2026-01-15T11:30:00.000Z"
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "memberships"
        ]
      },
      "post": {
        "operationId": "MembershipsController_create",
        "summary": "Join a community (create membership)",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC: admin | community_manager | moderator | user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMembershipDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": 5,
                    "userId": 3,
                    "communityId": 1,
                    "joinedAt": "2026-05-02T12:00:00.000Z"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "memberships"
        ]
      }
    },
    "/api/memberships/{id}": {
      "delete": {
        "operationId": "MembershipsController_remove",
        "summary": "Leave a community (delete membership by id)",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC: admin | community_manager | moderator | user",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Membership 5 removed"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "memberships"
        ]
      }
    },
...
  },
  "info": {
    "title": "Gameunity Backend API",
    "description": "In-memory NestJS backend for frontend integration and academic evaluation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Dev Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "x-role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-role",
        "description": "RBAC role header: admin | community_manager | moderator | user"
      }
    },
    "schemas": {
...
      "EventDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "Friday Scrim Night"
          },
          "description": {
            "type": "string",
            "example": "Weekly custom matches"
          },
          "communityId": {
            "type": "number",
            "example": 1
          },
          "date": {
            "type": "string",
            "example": "2026-05-09"
          },
          "time": {
            "type": "string",
            "example": "6:00 PM"
          },
          "type": {
            "type": "string",
            "example": "tournament"
          },
          "attendees": {
            "type": "number",
            "example": 48
          },
          "maxAttendees": {
            "type": "number",
            "example": 100
          },
          "status": {
            "type": "string",
            "example": "approved"
          },
          "createdBy": {
            "type": "string",
            "example": "Rajat"
          },
          "organiserId": {
            "type": "number",
            "example": 6
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "communityId",
          "date"
        ]
      },
      "CreateEventDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 3,
            "maxLength": 60,
            "example": "Clan Tournament Qualifier"
          },
          "description": {
            "type": "string",
            "minLength": 10,
            "maxLength": 220,
            "example": "Qualifier round for monthly tournament"
          },
          "communityId": {
            "type": "number",
            "example": 1
          },
          "date": {
            "type": "string",
            "example": "2026-05-20"
          },
          "time": {
            "type": "string",
            "example": "6:00 PM"
          },
          "type": {
            "type": "string",
            "example": "tournament",
            "enum": [
              "tournament",
              "hackathon",
              "workshop",
              "meetup",
              "watchparty"
            ]
          },
          "attendees": {
            "type": "number",
            "example": 0
          },
          "maxAttendees": {
            "type": "number",
            "example": 100
          },
          "status": {
            "type": "string",
            "example": "pending",
            "enum": [
              "pending",
              "approved",
              "rejected",
              "cancelled"
            ]
          },
          "createdBy": {
            "type": "string",
            "example": "Rajat"
          },
          "organiserId": {
            "type": "number",
            "example": 6,
            "description": "Links to the organiser's user id — used for organizer-plan limit checks (createdBy is just a display string)"
          },
          "category": {
            "type": "string",
            "example": "Esports"
          },
          "coverImage": {
            "type": "string",
            "example": "data:image/png;base64,..."
          }
        },
        "required": [
          "title",
          "description",
          "communityId",
          "date"
        ]
      },
      "UpdateEventDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 3,
            "maxLength": 60,
            "example": "Clan Tournament Qualifier"
          },
          "description": {
            "type": "string",
            "minLength": 10,
            "maxLength": 220,
            "example": "Qualifier round for monthly tournament"
          },
          "communityId": {
            "type": "number",
            "example": 1
          },
          "date": {
            "type": "string",
            "example": "2026-05-20"
          },
          "time": {
            "type": "string",
            "example": "6:00 PM"
          },
          "type": {
            "type": "string",
            "example": "tournament",
            "enum": [
              "tournament",
              "hackathon",
              "workshop",
              "meetup",
              "watchparty"
            ]
          },
          "attendees": {
            "type": "number",
            "example": 0
          },
          "maxAttendees": {
            "type": "number",
            "example": 100
          },
          "status": {
            "type": "string",
            "example": "pending",
            "enum": [
              "pending",
              "approved",
              "rejected",
              "cancelled"
            ]
          },
          "createdBy": {
            "type": "string",
            "example": "Rajat"
          },
          "organiserId": {
            "type": "number",
            "example": 6,
            "description": "Links to the organiser's user id — used for organizer-plan limit checks (createdBy is just a display string)"
          },
          "category": {
            "type": "string",
            "example": "Esports"
          },
          "coverImage": {
            "type": "string",
            "example": "data:image/png;base64,..."
          }
        }
      },
...
      "CreateMembershipDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number",
            "example": 3,
            "description": "ID of the user joining the community"
          },
          "communityId": {
            "type": "number",
            "example": 1,
            "description": "ID of the community to join"
          }
        },
        "required": [
          "userId",
          "communityId"
        ]
      },
...
    }
  }
}